feat: use libadwaita-like favicon
Switches the favicon from the GNOME icon to a libadwaita-like icon, as the
website lists apps that are not specifically made for GNOME, but still
utilizing libadwaita. Additionally, this avoid confusing, about the
associations with GNOME.
The icon was originally used by the updated GNOME websites, so credit to Jakub
Steiner.

Ref: https://gitlab.gnome.org/Teams/Websites/www.gnome.org/-/merge_requests/23